refactor(scripting): extract Lua event bus logic and global event functions to event_bus.rs
